feat: purchases data safety, OCR scan, strict decimals
- Enforce duplicate trade purchases on save (409 + force_duplicate bypass); soft-delete
- Expose has_missing_details; Hive-backed wizard draft + post-save banners/reminders
- Bill scan endpoints and Flutter wizard integration; ledger/history UX
- Strict decimal migrations and StrictDecimal-aligned purchase math

+++ backend/app/models/catalog.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import datetime, timezone
+from decimal import Decimal
 
 from sqlalchemy import DateTime, ForeignKey, Integer, Numeric, String, UniqueConstraint, Uuid
 from sqlalchemy.orm import Mapped, mapped_column, relationship
@@ -53,18 +54,18 @@ class CatalogItem(Base):
     )
     name: Mapped[str] = mapped_column(String(512))
     default_unit: Mapped[str | None] = mapped_column(String(32), nullable=True)
-    default_kg_per_bag: Mapped[float | None] = mapped_column(Numeric(18, 4), nullable=True)
-    default_items_per_box: Mapped[float | None] = mapped_column(Numeric(18, 4), nullable=True)
-    default_weight_per_tin: Mapped[float | None] = mapped_column(Numeric(18, 4), nullable=True)
+    default_kg_per_bag: Mapped[Decimal | None] = mapped_column(Numeric(12, 3), nullable=True)
+    default_items_per_box: Mapped[Decimal | None] = mapped_column(Numeric(12, 3), nullable=True)
+    default_weight_per_tin: Mapped[Decimal | None] = mapped_column(Numeric(12, 3), nullable=True)
     hsn_code: Mapped[str | None] = mapped_column(String(32), nullable=True)
     # Internal / ERP product code (optional; matches seed JSON "code").
     item_code: Mapped[str | None] = mapped_column(String(64), nullable=True)
-    tax_percent: Mapped[float | None] = mapped_column(Numeric(18, 4), nullable=True)
-    default_landing_cost: Mapped[float | None] = mapped_column(Numeric(18, 4), nullable=True)
-    default_selling_cost: Mapped[float | None] = mapped_column(Numeric(18, 4), nullable=True)
+    tax_percent: Mapped[Decimal | None] = mapped_column(Numeric(5, 2), nullable=True)
+    default_landing_cost: Mapped[Decimal | None] = mapped_column(Numeric(12, 2), nullable=True)
+    default_selling_cost: Mapped[Decimal | None] = mapped_column(Numeric(12, 2), nullable=True)
     default_purchase_unit: Mapped[str | None] = mapped_column(String(32), nullable=True)
     default_sale_unit: Mapped[str | None] = mapped_column(String(32), nullable=True)
-    last_purchase_price: Mapped[float | None] = mapped_column(Numeric(18, 4), nullable=True)
+    last_purchase_price: Mapped[Decimal | None] = mapped_column(Numeric(12, 2), nullable=True)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), default=utcnow)
 
     category = relationship("ItemCategory", back_populates="items")
@@ -131,7 +132,7 @@ class CatalogVariant(Base):
         Uuid(as_uuid=True), ForeignKey("catalog_items.id"), index=True
     )
     name: Mapped[str] = mapped_column(String(512))
-    default_kg_per_bag: Mapped[float | None] = mapped_column(Numeric(18, 4), nullable=True)
+    default_kg_per_bag: Mapped[Decimal | None] = mapped_column(Numeric(12, 3), nullable=True)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), default=utcnow)
 
     item = relationship("CatalogItem", back_populates="variants")
